Score and Rerank support (#5)
This adds support for score and rerank requests using the `/v1/score` and `/v1/rerank` endpoints, respectively. These are supported by vLLM and Parasail, but not OpenAI. Rerank is similar to Cohere's rerank.

Score workload is automatically detected by looking for the text_1 and text_2 fields. Rerank looks for query and documents.